Resolve #9: FIND a Task
A simple Junit Test was created. Also, the wrapper classes from 'deleteById' and 'findById', at TaskDao interface, were changed from Integer to Long

+++ src/main/java/model/dao/TaskDao.java
@@ -25,12 +25,12 @@ public interface TaskDao {
 	/**
 	 * @param id
 	 */
-	void deletedById(Integer id);
+	void deletedById(Long id);
 	
 	/**
 	 * @param id
 	 */
-	Task findById(Integer id);
+	Task findById(Long id);
 	
 	/**
 	 * @return List<Task>
